requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
click==8.2.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.18.2
mypy_extensions==1.1.0
passlib==1.7.4
pathspec==0.12.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
ruff==0.14.1
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
WTForms==3.1.2
